# complete/openapi.yaml
openapi: 3.0.1
info:
  title: Simple Social Media API
  description: A basic Social Networking Service (SNS) API that allows users to create, retrieve, update, and delete posts; add comments; and like/unlike posts.
  version: 1.0.0
  contact:
    name: Contoso Product Team
    email: support@contoso.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
  /posts:
    get:
      summary: List all posts
      description: Retrieve all recent posts to browse what others are sharing
      operationId: getPosts
      tags:
        - Posts
      responses:
        "200":
          description: Successfully retrieved posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new post
      description: Create a new post to share something with others
      operationId: createPost
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPostRequest"
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  schemas:
    Post:
      type: object
      required:
        - id
        - username
        - content
        - createdAt
        - updatedAt
        - likesCount
        - commentsCount
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the post
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          minLength: 1
          maxLength: 50
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          minLength: 1
          maxLength: 2000
          description: Content of the post
          example: "Just had an amazing hike in the mountains! #outdoorlife"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the post was created
          example: "2025-06-01T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the post was last updated
          example: "2025-06-01T10:30:00Z"
        likesCount:
          type: integer
          minimum: 0
          description: Number of likes on the post
          example: 15
        commentsCount:
          type: integer
          minimum: 0
          description: Number of comments on the post
          example: 3
...
    NewPostRequest:
      type: object
      required:
        - username
        - content
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          minLength: 1
          maxLength: 2000
          description: Content of the post
          example: "Just had an amazing hike in the mountains! #outdoorlife"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code or type
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "The request body is invalid"
        details:
          type: array
          description: Additional error details (optional)
          items:
            type: string
          example: ["username is required", "content must not be empty"]
  responses:
    BadRequest:
      description: Bad request - invalid input or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "VALIDATION_ERROR"
            message: "The request body is invalid"
            details: ["username is required", "content must not be empty"]
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "INTERNAL_ERROR"
            message: "An unexpected error occurred"
...